split out data files
